Change the type of time_t to long
Change the type of time_t from uint32_t to long

+++ os/include/time.h
@@ -142,7 +142,7 @@
  * Public Types
  ********************************************************************************/
 
-typedef uint32_t time_t;		/* Holds time in seconds */
+typedef long  time_t;		/* Holds time in seconds */
 typedef uint8_t clockid_t;		/* Identifies one time base source */
 typedef FAR void *timer_t;		/* Represents one POSIX timer */
